Aggregate environment-related logic to the common class `Environment.cs`.
