+++ HouseRules_Configuration/ConfigurationMod.cs
@@ -3,15 +3,14 @@
     using System;
     using System.Collections.Generic;
     using Common;
+    using Common.UI;
     using HouseRules.Configuration.UI;
     using MelonLoader;
     using UnityEngine;
 
     internal class ConfigurationMod : MelonMod
     {
-        private const string DemeoPCEditionString = "Demeo PC Edition";
         private const int LobbySceneIndex = 1;
-        private const int VrHangoutsSceneIndex = 43;
 
         internal static readonly MelonLogger.Instance Logger = new MelonLogger.Instance("HouseRules:Configuration");
         internal static readonly ConfigManager ConfigManager = ConfigManager.NewInstance();
@@ -53,7 +52,7 @@
 
         public override void OnSceneWasInitialized(int buildIndex, string sceneName)
         {
-            if (MelonUtils.CurrentGameAttribute.Name == DemeoPCEditionString)
+            if (Environments.IsPcEdition())
             {
                 if (buildIndex != LobbySceneIndex)
                 {
@@ -65,7 +64,7 @@
                 return;
             }
 
-            if (buildIndex == VrHangoutsSceneIndex)
+            if (Environments.IsInHangouts())
             {
                 Logger.Msg("Recognized lobby in Hangouts. Loading UI.");
                 _ = new GameObject("HouseRulesUiHangouts", typeof(HouseRulesUiHangouts));
@@ -102,7 +101,8 @@
                 catch (Exception e)
                 {
                     FailedRulesetFiles.Add(file);
-                    Logger.Warning($"Failed to import and register ruleset from file [{file}]. Skipping that ruleset: {e}");
+                    Logger.Warning(
+                        $"Failed to import and register ruleset from file [{file}]. Skipping that ruleset: {e}");
                 }
             }
         }
